requirements.txt
astroid==2.3.3
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
cryptography==2.8
cx-Freeze==6.1
decorator==4.4.2
future==0.18.2
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.0.0
pycparser==2.20
pylint==2.4.4
PyMySQL==0.9.3
python-telegram-bot==12.4.2
requests==2.23.0
six==1.14.0
tornado==6.0.4
typed-ast==1.4.1
urllib3==1.25.8
wrapt==1.11.2
